requirements.txt
yfinance>=0.2.30
numpy>=1.24
pandas>=2.0
gspread>=6.0
google-auth>=2.20
